edit Week1 step3
